requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
numpy
matplotlib
# ecdsa
# paramiko
# pycrypto
